+++ include/simpleble/Peripheral.h
@@ -1,13 +1,24 @@
 #pragma once
 
+#include <memory>
+
+#include <simpleble/Exceptions.h>
+#include <simpleble/Types.h>
+
 namespace SimpleBLE {
+
+class PeripheralBase;
 
 class Peripheral {
   public:
     Peripheral();
     ~Peripheral();
 
-  private:
+    std::string identifier();
+    BluetoothAddress address();
+
+  protected:
+    std::shared_ptr<PeripheralBase> internal_;
     
 };