+++ scripts/macos/build_lib.sh
@@ -14,7 +14,7 @@ realpath() {
 }
 
 PROJECT_ROOT=$(realpath $(dirname `realpath $0`)/../..)
-BUILD_PATH=$PROJECT_ROOT/.build
+BUILD_PATH=$PROJECT_ROOT/build
 
 # Parse the received commands
 while :; do