+++ src/linux/PeripheralBase.cpp
@@ -57,13 +57,7 @@ bool PeripheralBase::is_connected() {
     return device->Property_Connected() && device->Property_ServicesResolved();
 }
 
-bool PeripheralBase::is_connectable() {
-    // TODO its a stub, replace with real implementation
-    if (is_connected()) {
-        return false;
-    }
-    return true;
-}
+bool PeripheralBase::is_connectable() { return identifier_ != ""; }
 
 std::vector<BluetoothService> PeripheralBase::services() {
     auto device = _get_device();
@@ -80,19 +74,22 @@ std::vector<BluetoothService> PeripheralBase::services() {
 }
 
 std::map<uint16_t, ByteArray> PeripheralBase::manufacturer_data() {
-    // TODO: Implement
-    return {};
+    auto device = _get_device();
+
+    std::map<uint16_t, ByteArray> manufacturer_data;
+    for (auto& [manufacturer_id, value_array] : device->get_manufacturer_data()) {
+        manufacturer_data[manufacturer_id] = ByteArray((const char*)value_array.data(), value_array.size());
+    }
+
+    return manufacturer_data;
 }
 
 ByteArray PeripheralBase::read(BluetoothUUID service, BluetoothUUID characteristic) {
     // TODO: Check if the characteristic is readable.
     auto characteristic_object = _get_characteristic(service, characteristic);
 
-    // FIXME: SimpleDBus does not have this feature yet.
-    // std::vector<uint8_t> value = characteristic_object->get_value();
-    // return ByteArray((const char*)value.data(), value.size());
-
-    return ByteArray();
+    std::vector<uint8_t> value = characteristic_object->Property_Value();
+    return ByteArray((const char*)value.data(), value.size());
 }
 
 void PeripheralBase::write_request(BluetoothUUID service, BluetoothUUID characteristic, ByteArray data) {