+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimalist Linux implementation to run the `list_adapters` example.
 - Minimalist Windows implementation to run the `list_adapters` example.
 - Minimalist MacOS implementation to run the `list_adapters` example.
+- Minimalist Linux implementation to run the `scan` example.