+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.2] - 2021-XX-XX
 ### Added
-- 
+- Safe implementation of `SimpleBLE::Adapter` and `SimpleBLE::Peripheral` classes.
 ### Fixed
 - Compilation errors that came up during development. _(Thanks fidoriel!)_