+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.2] - 2021-XX-XX
+## [0.0.2] - 2021-10-09
 
 ### Added
 - Safe implementation of `SimpleBLE::Adapter` and `SimpleBLE::Peripheral` classes.
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - WinRT buffer allocation would fail. _(Thanks PatrykSajdok!)_
 - `SimpleBLE::Adapter` would fail to stop scanning. _(Thanks PatrykSajdok!)_
 - Switched WinRT initialization to single-threaded.
+
+### Changed
+- SimpleDBus dependency not migrated to OpenBluetoothToolbox.
 
 ## [0.0.1] - 2021-09-06